+++ packages/utils/test/index.test.ts
@@ -1,73 +1,5 @@
 import { describe, test, expect, assert } from "vitest";
-import { createComputed, createRoot } from "solid-js";
-import {
-  createStaticStore,
-  handleDiffArray,
-  arrayEquals,
-  createHydratableSignal,
-  createHydratableStaticStore,
-} from "../src";
-
-describe("createStaticStore", () => {
-  test("individual keys only update when changed", () => {
-    createRoot(dispose => {
-      const _shape = { a: 1, b: 2, c: 3, d: [0, 1, 2] };
-      const [state, setState] = createStaticStore(_shape);
-
-      expect(state).toEqual(_shape);
-      expect(_shape, "original input shouldn't be mutated").toEqual({
-        a: 1,
-        b: 2,
-        c: 3,
-        d: [0, 1, 2],
-      });
-
-      setState({ a: 9, d: [3, 2, 1] });
-
-      expect(state).toEqual({ a: 9, b: 2, c: 3, d: [3, 2, 1] });
-      expect(_shape, "original input shouldn't be mutated").toEqual({
-        a: 1,
-        b: 2,
-        c: 3,
-        d: [0, 1, 2],
-      });
-
-      let aUpdates = -1;
-      createComputed(() => {
-        state.a;
-        aUpdates++;
-      });
-      expect(aUpdates).toBe(0);
-
-      setState({
-        b: 3,
-      });
-      expect(aUpdates).toBe(0);
-      setState("a", 4);
-      expect(aUpdates).toBe(1);
-
-      dispose();
-    });
-  });
-
-  // tss("able to listen to key, not yet added", () =>
-  //   createRoot(dispose => {
-  //     const _shape = {};
-  //     const [state, setState] = createShallowStore<{ a?: number }>(_shape);
-
-  //     let captured: any[] = [];
-  //     createComputed(() => captured.push(state.a));
-
-  //     assert.equal(captured, [undefined]);
-
-  //     setState("a", 1);
-
-  //     assert.equal(captured, [undefined, 1]);
-
-  //     dispose();
-  //   })
-  // );
-});
+import { handleDiffArray, arrayEquals, createHydratableSignal } from "../src";
 
 describe("handleDiffArray", () => {
   test("handleAdded called for new array", () => {
@@ -167,16 +99,6 @@ describe("createHydratableSignal", () => {
   test("createHydratableSignal() - CSR", () => {
     const [state, setState] = createHydratableSignal("server", () => "client");
     expect(state()).toEqual("client");
-    expect(setState).toBeInstanceOf(Function);
-  });
-});
-
-describe("createHydratableStaticStore", () => {
-  test("createHydratableStaticStore() - CSR", () => {
-    const [state, setState] = createHydratableStaticStore({ foo: "server" }, () => ({
-      foo: "client",
-    }));
-    expect(state).toEqual({ foo: "client" });
     expect(setState).toBeInstanceOf(Function);
   });
 });