+++ packages/connectivity/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @solid-primitives/connectivity
 
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+  - @solid-primitives/event-listener@2.2.10
+  - @solid-primitives/rootless@1.3.2
+
 ## 0.3.11
 
 ### Patch Changes