+++ packages/controlled-props/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/controlled-props
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+
 ## 0.0.8
 
 ### Patch Changes