+++ packages/mutation-observer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/mutation-observer
 
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+
 ## 1.1.9
 
 ### Patch Changes