+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @solid-primitives/styles
 
+## 0.0.108
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+  - @solid-primitives/rootless@1.3.2
+
 ## 0.0.107
 
 ### Patch Changes